requirements.txt
numpy
pandas
xlsxwriter
streamlit
plotly-express
stqdm
top2vec
top2vec[sentence_transformers]
top2vec[sentence_encoders]
top2vec[indexing]
spacy>=3.0.0,<4.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz#egg=en_core_web_sm